remove ktos scoop bucket; fix transformer tests (accordingly with latest version; add a check to test ps1 return code to fail on failing test cases

+++ test.ps1
@@ -12,6 +12,10 @@ catch {
 }
 Pop-Location
 
+if ($lastexitcode -ne 0) {
+    throw ("Exec: " + $errorMessage)
+}
+
 Push-Location cmake\test\common.cmake\
 if (Test-Path .cmaketest) {
     Remove-Item .cmaketest -Recurse -Force
